(partially) cleaned up version of the EA approach + simulator
This contains:

- config.ini - for a lot of settings, many of them experimental

- our insulin/diabetes SIMULATOR, including code to read data files (no data included though).  The simulator implements a variant on the Bergman model, together with Breton's adaptations for activity (for which the heart rate is a proxy)

- a finite state machine to SIMULATE HEART RATES of patients over a timespan of many days. The FSM uses the (randomly generated, based on probabilities from the config.ini) events (like eating and injecting insulin) to generate a somewhat realistic heart rate. The FSM uses states like 'asleep', 'walking', 'sitting', 'running' etc...

- MACHINE LEARNING, specifically a subpopulation based evolutionary algorithm (without crossover), sometimes still called particle filter in the code, because that is how it started, and it evolved into a no-crossover-hierarchical EA.
The core of the algorithm always is: guess (in some smart way, using particle filter, EA or whatever) the Bergman and Breton parameters for the patient, then simulate a virtual patient for these parameters and evaluate how good the match is. The current EA approximation has populations of 'particles' (legacy of particle filter :-) that have values for the bergman+breton parameters. Particles that score better at simulating the "real' patient, get higher fitness (lower error).
The EA has several 'sub populations', each of which evolve independently thus preventing a convergence on one single (probably suboptimal) solution. Some other tricks and ideas are included as well, such as using a hash table like structure to detect if a sub population is currently 'investigating' an erea of the solution space that has already been covered. If so, this sub population can be 'disbanded' and reinstated elsewhere in the solution space. The same will happen for subpopulations that 'stagnate' (don't increase their performance over some time, no lower error rates). The top level EA keeps track of the sub populations, the best performance, etc..
The learning is done in phases. All the 'particles' (entities?) from the sub populations train on a piece of the timeline. Then the best of each is is tested on a longer piece of the time line, to prevent overfitting). Then the best particle of all is used to do the final predictions, and is compared to the real (or simulated) data on an even longer timescale.
This is all very much experimental. There was not enough time to arrive at a conclusive form for the algorithm during the project :-(  It seems that on *real patient data* the algorithm still gets stuck in a suboptimum. Further investigation is needed into ways to get it out of  this suboptimum. Possible venues could include: using other metrics for calculating fitness (error). Curently a weighted RMSE is used, and some small experiments were done using the shape of the 'real' curve and the ML predicted one (by looking at the derivative etc..). Other options might be some way to backpropagate information back into the differential equations of Bergman and Breton. Currently, the particles have parameters for these equations, the equations are run (a virtual patient is simulated for a time span, based on known events like eating, bolus, heart rate) and the output is compared with the 'real' data. But the output can only be compared on the final result, the blood glucose curve. All other steps in the differential equations (active insulin, gamma, X, etc..) are hypothetical and cannot be measured. So the diff eq. are 'one way'. You start with hidden parameters + equations and end up with BG. Currently the difference between the 'real' BG (measured or simulated) and the 'machine learning' BG is the only thing that drives the evolution. Bigger difference (RMSE, weighted) is lower fitness. Next iteration: try again with the highest scoring part of the population (+random mutation). But what if we can backpropagate (just like in Neural networks) the difference (error signal) into the diff. eq? I haven't found a way yet, but maybe it is possible to use the error signal to make more directed adjustments to the parameters 'up' in the chain of diff. eq.. I have tried to work this out on paper analogous to backprop in artificial neural networks, but got nowhere. But there may still be an option here. Things like: if the 'machine learning predicted' BG is too low in general, just increase the 'base' glucose parameter in Bergman eq...  But there are about a dozen other parameters in Bergman +Breton that don't have such an obvious fix, because the combination of e.g. p1, p2, p3 and p4 gives rise to particular shapes of the BG curve.

- LIBSHARPTAVE (slightly modified version, based on open source project). I used octave (open source version of Matlab) to do my plotting, so the C# code generates .m files. The LibSharpTave code provides a means to execute an OCTAVE command from within the C# code.